+++ core/src/main/java/com/bitplay/persistance/repository/GuiParamsRepository.java
@@ -1,12 +1,11 @@
 package com.bitplay.persistance.repository;
 
 import com.bitplay.persistance.domain.AbstractParams;
-import java.math.BigInteger;
 import org.springframework.data.repository.CrudRepository;
 
 /**
  * Created by Sergey Shurmin on 6/17/17.
  */
-public interface GuiParamsRepository extends CrudRepository<AbstractParams, BigInteger> {
+public interface GuiParamsRepository extends CrudRepository<AbstractParams, Long> {
 
 }
